api_meli/requirements.txt
Django==4.0.6
django-debug-toolbar==3.2.4
django-dotenv==1.4.2
django-extensions==3.1.5
django-filter==21.1
django-guardian==2.4.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
Pillow==9.1.0
Werkzeug==2.0.3
drf-nested-routers==0.93.4
django-cors-headers==3.11.0
coverage==6.3.2
pydantic==1.9.0
whitenoise==6.2.0
gunicorn
google-api-python-client==1.7.9
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
Faker==13.14.0
mock==4.0.3
flake8==5.0.0